warnings

+++ scilab/modules/graphics/sci_gateway/c/sci_xget.c
@@ -142,12 +142,12 @@ int sci_xget(char *fname,unsigned long fname_len)
     sciPointObj *psubwin = sciGetCurrentSubWin();
     x2=0;
     if(strcmp(cstk(l1),"font")==0){
-      x1[0] = sciGetFontStyle(psubwin);
-      x1[1] = sciGetFontSize(psubwin);
+      x1[0] = (integer)sciGetFontStyle(psubwin);
+      x1[1] = (integer)sciGetFontSize(psubwin);
       x2 = 2;
     }
     else if(strcmp(cstk(l1),"font size")==0){
-      x1[0] = sciGetFontSize(psubwin);
+      x1[0] = (integer)sciGetFontSize(psubwin);
       x2 = 1;
     }
     else if(strcmp(cstk(l1),"dashes")==0){
